+++ CHANGELOG.md
@@ -3,7 +3,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased
 ### Changed
-- Flags are listed in alphabetical order.
+- Commands and flags are now listed in alphabetical order.
 
 ## 0.1.5 (2017-10-26)
 ### Changed